Optimize OrderService
